// src/app/common-methods/models/email-config.ts
export class EmailConfig {
  emailConfigId: number;
  applicationId: number;
  applicationName: string;
  emailGroupId: number;
  emailGroupName: string;
  emailKeyId: number;
  emailKeyValue: string;
  emailDestType: number;
  emailDest: string;
  emailAddress: string;
  active: boolean;
  logginId: number;
}